move feed2room to own file
